refactor: extract scan and watch commands into phi_scan/cli/scan.py
Final slice of the CLI monolith split. Moves scan orchestration
(sequential/parallel progress dispatch, audit persistence, notification
dispatch, CI integration dispatch, framework flag resolution) and the
watch command out of phi_scan/cli/__init__.py into a dedicated module.

Also addresses deferred cleanup called out in the plan doc:
- Renames _prepare_scan_phase to _collect_scan_targets_for_phase to
  satisfy the verb-noun naming rule.
- Relocates the orphan _SPINNER_NOTIFY_MESSAGE and
  _SPINNER_CONFIG_LOAD_MESSAGE constants to scan.py alongside their
  only callers.

phi_scan/cli/__init__.py is now a 128-line thin dispatcher (down from
1492 at the start of the pass). No behavior change; help output and
command set are unchanged.
